requirements.txt
## The following requirements were added by pip --freeze:
Django==1.7.1
Markdown==2.5.2
Mezzanine==3.1.10
Pillow==2.6.1
argparse==1.2.1
beautifulsoup4==4.1.3
bleach==1.4
django-appconf==0.6
django-braces==1.4.0
django-compressor==1.4
django-filter==0.8
django-model-utils==2.2
django-oauth-toolkit==0.7.2
djangorestframework==2.4.4
envoy==0.0.3
filebrowser-safe==0.3.6
future==0.9.0
futures==2.2.0
grappelli-safe==0.3.13
gunicorn==19.1.1
html5lib==0.999
oauth2-provider==0.0
oauthlib==0.7.2
pep8==1.5.7
psycopg2==2.5.4
pyflakes==0.8.1
python-memcached==1.53
python-swiftclient==2.3.1
pytz==2014.9
requests==2.4.3
requests-oauthlib==0.4.2
setproctitle==1.1.8
simplejson==3.6.5
six==1.8.0
tzlocal==1.0
wsgiref==0.1.2
